test(djf.io): co-locate e2e tests, fix CI lint failures
Move e2e specs from tests/e2e/*.spec.ts into src/pages/*.e2e.test.ts
next to the routes they exercise. Files in src/pages/ are prefixed
with `_` so Astro's filesystem router skips them (built-in escape
hatch — Astro has no other exclude config). Playwright's testMatch
broadens to **/*.e2e.test.ts so the same convention applies anywhere
in src/.

Address pre-existing lint failures the new CI workflow surfaced:

- .oxlintrc.json: disable triple-slash-reference for **/*.d.ts
  (Astro's auto-generated env.d.ts uses the idiom)
- apps/djf.io/biome.json: disable noUnusedImports/noUnusedVariables
  for **/*.astro — biome 2's astro parser doesn't see template usage
- apps/djf.io/.vscode/settings.json: missing comma broke biome's
  JSON parser
- biome --write reformats in src/content/config.ts,
  src/pages/blog/index.astro, src/pages/rss.xml.ts (trailing commas
  + organize-imports per existing biome.jsonc rules)
