+++ apps/djf.io/.vscode/settings.json
@@ -1,32 +1,32 @@
 {
-	"cSpell.words": [
-		"Angelhack",
-		"architecting",
-		"astro",
-		"astrojs",
-		"Atumsoft",
-		"commenters",
-		"DALL-E",
-		"DEVBOX",
+  "cSpell.words": [
+    "Angelhack",
+    "architecting",
+    "astro",
+    "astrojs",
+    "Atumsoft",
+    "commenters",
+    "DALL-E",
+    "DEVBOX",
     "davidjfelix",
-		"doomer",
-		"gatekept",
-		"Grammarly",
-		"Hackathon",
-		"half-assed",
-		"headshot",
-		"Instameals",
-		"iotv",
-		"pandacss",
-		"Pepperonio",
+    "doomer",
+    "gatekept",
+    "Grammarly",
+    "Hackathon",
+    "half-assed",
+    "headshot",
+    "Instameals",
+    "iotv",
+    "pandacss",
+    "Pepperonio",
     "posthog",
-		"Strava",
-		"struct",
-		"Swarmfarm",
-		"tsconfigs",
-		"VNDLY"
-	],
-	"cSpell.enableFiletypes": ["astro", "just", "mdx", "nix"]
+    "Strava",
+    "struct",
+    "Swarmfarm",
+    "tsconfigs",
+    "VNDLY"
+  ],
+  "cSpell.enableFiletypes": ["astro", "just", "mdx", "nix"],
   "[css]": {
     "editor.defaultFormatter": "biomejs.biome"
   },
