+++ apps/djf.io/src/pages/blog/index.astro
@@ -4,7 +4,7 @@ import {css} from 'styled-system/css'
 import BaseLayout from '../../layouts/BaseLayout.astro'
 
 const posts = (await getCollection('blog')).sort(
-  (a, b) => new Date(b.data.date).valueOf() - new Date(a.data.date).valueOf()
+  (a, b) => new Date(b.data.date).valueOf() - new Date(a.data.date).valueOf(),
 )
 ---
