+++ apps/djf.io/src/pages/rss.xml.ts
@@ -1,12 +1,10 @@
-import rss from '@astrojs/rss'
 import {getCollection} from 'astro:content'
+import rss from '@astrojs/rss'
 import type {APIContext} from 'astro'
 
 export async function GET(context: APIContext) {
   const posts = await getCollection('blog')
-  const sortedPosts = posts.sort(
-    (a, b) => b.data.date.valueOf() - a.data.date.valueOf()
-  )
+  const sortedPosts = posts.sort((a, b) => b.data.date.valueOf() - a.data.date.valueOf())
 
   return rss({
     title: "David J. Felix's Blog",
