+++ docs/projects/djf-io-testing/2026-04-29-progress.md
@@ -3,24 +3,37 @@
 ## Done
 
 - Added `@playwright/test` to `apps/djf.io` devDeps
-- `playwright.config.ts` with single Chromium project, `webServer` invoking `astro preview`, CI-aware retries/workers/reporters
-- `tests/e2e/`:
-  - `home.spec.ts` — landing heading, tagline, blog link
-  - `blog.spec.ts` — blog index ordering, MDX post body + headings, tags index, tag filter
-  - `rss.spec.ts` — RSS feed XML and entries
+- `playwright.config.ts`: single Chromium project, `webServer` invoking `astro preview`, `testDir: './src'`, `testMatch: '**/*.e2e.test.ts'`, CI-aware retries/workers/reporters
+- E2E specs co-located next to the source they exercise:
+  - `src/pages/_index.e2e.test.ts`
+  - `src/pages/blog/_index.e2e.test.ts`
+  - `src/pages/blog/_[...slug].e2e.test.ts`
+  - `src/pages/blog/tags/_index.e2e.test.ts`
+  - `src/pages/blog/tags/_[tag].e2e.test.ts`
+  - `src/pages/_rss.xml.e2e.test.ts`
 - `package.json` scripts: `test`, `test:e2e`, `test:e2e:ui`
 - `.gitignore`: Playwright artifacts (`test-results/`, `playwright-report/`, `blob-report/`, `playwright/.cache/`)
 - `.github/workflows/djf-io-ci.yml` — two jobs (lint+build, e2e); path-filtered; uses `jdx/mise-action` so biome/oxlint/node/pnpm match `.config/mise.toml`; Playwright browsers cached by version
 
+## Lint debt addressed (PR was first to actually run `pnpm lint` in CI)
+
+- `.oxlintrc.json`: override `typescript/triple-slash-reference` off for `**/*.d.ts` (Astro's `env.d.ts` uses the idiom)
+- `apps/djf.io/biome.json`: override to disable `noUnusedImports` / `noUnusedVariables` for `**/*.astro` — biome 2's `.astro` parser doesn't see template usage and produces false positives
+- `.vscode/settings.json`: fixed missing comma that broke biome's JSON parser
+- `biome check --write .` reformat in `src/content/config.ts`, `src/pages/blog/index.astro`, `src/pages/rss.xml.ts` (trailing-comma + organize-imports per the existing biome.jsonc rules)
+
 ## Validated locally
 
-- `pnpm install`, `pnpm build`, `pnpm exec playwright install chromium`, `pnpm test:e2e` — all 7 tests pass (~3.6s)
-- Biome (via npx) clean on new files
+- `pnpm build`, `pnpm test:e2e` — green (7 tests, ~14s)
+- `npx @biomejs/biome@2 check .` — exit 0
+- `npx oxlint@1 .` — exit 0
 
 ## Decisions
 
-- **Vitest deferred**: no shared/util code in djf.io worth unit-testing yet. Plan documents the trigger to add it later. Avoids fabricating helpers solely for tests.
-- **Single browser (Chromium)**: blog is server-rendered HTML; cross-browser coverage adds CI minutes for marginal gain on this app. Easy to expand.
+- **Tests co-located, named `xxx.e2e.test.ts`** next to the source under test. Unit-test counterpart will be `xxx.test.ts` when Vitest lands.
+- **`_` prefix on test files inside `src/pages/`**: Astro's filesystem router treats every `.ts` in `src/pages/` as an endpoint and gives no exclude config. The leading underscore is Astro's built-in escape hatch (any file/dir starting with `_` is skipped). Outside `src/pages/` the prefix isn't needed.
+- **Vitest deferred**: no shared/util code in djf.io worth unit-testing yet. Plan documents the trigger to add it later.
+- **Single browser (Chromium)**: blog is server-rendered HTML; cross-browser coverage adds CI minutes for marginal gain on this app.
 - **`retries: 2` in CI**: standard Playwright practice for flake mitigation; locally `0`.
 
 ## Next
